Buttons Work, Bug Fixes, and Refactored
Fixed bug that was crashing the game if you resized before clicking new deal. Fixed bug preventing pause button to work. All buttons are disabled except for new game at the start. Then restart and pause are enabled. Undo and Redo are only enabled if there is anything to undo or redo respectively. Refactored trans module to acts. Refactored redeal to restart to be consistent throughout the program.  Fixed bug that allowed undo of picking up a card and placing it back onto the same tile.