Winnable Hands, Side Bars, and Checkbox
Created a Shuffler class that will give a winnable hand or a random hand. The winnable hand is rudimentary at the moment. Updated the sidebar control to be able to be on any side of the screen, created a better demo in controls to test everything, created a checkbox control, and added a right side bar with a checkbox to control if you want winnable hands or not.