GUI Fixes
The slider bar works, and contains ruler lines. Sidebar is moved to its own file, and tested there with less controls. Controls has its own test that tests all the controls. Refresh sidebar event was changed to refresh ui to be used to generally refresh any ui that contains a control. Fixed border rect, so that its redal size and position is represented by the usual variables. Updated the back color of stop watch and counter box to a softer color.